notion-generator/requirements.txt
streamlit>=1.35.0
anthropic>=0.50.0
notion-client>=2.2.1
python-dotenv>=1.0.1
